# feat: ewma use p2c to improve performance