--- a/apisix/balancer/ewma.lua
+++ b/apisix/balancer/ewma.lua
@@ -3,7 +3,6 @@
 -- Inspiration drawn from:
 -- https://github.com/twitter/finagle/blob/1bc837c4feafc0096e43c0e98516a8e1c50c4421
 --   /finagle-core/src/main/scala/com/twitter/finagle/loadbalancer/PeakEwma.scala
-
 local core = require("apisix.core")
 local resty_lock = require("resty.lock")
 
@@ -12,111 +11,108 @@ local ngx_shared = ngx.shared
 local ngx_now = ngx.now
 local math = math
 local pairs = pairs
-local tostring = tostring
-local string = string
 
 local DECAY_TIME = 10 -- this value is in seconds
 local LOCK_KEY = ":ewma_key"
 
 local shm_ewma = ngx_shared.balancer_ewma
-local shm_last_touched_at= ngx_shared.balancer_ewma_last_touched_at
+local shm_last_touched_at = ngx_shared.balancer_ewma_last_touched_at
 
-local lrucache_addr = core.lrucache.new({
-    ttl = 300, count = 1024
-})
-local lrucache_trans_format = core.lrucache.new({
-    ttl = 300, count = 256
-})
+local lrucache_addr = core.lrucache.new({ttl = 300, count = 1024})
+local lrucache_trans_format = core.lrucache.new({ttl = 300, count = 256})
 
-local ewma_lock, ewma_lock_err = resty_lock:new("balancer_ewma_locks", {timeout = 0, exptime = 0.1})
+local ewma_lock, ewma_lock_err = resty_lock:new("balancer_ewma_locks",
+                                                {timeout = 0, exptime = 0.1})
 if not ewma_lock then
-  error(ewma_lock_err)
+    error(ewma_lock_err)
 end
 
-local _M = { name = "ewma" }
+local _M = {name = "ewma"}
 
 local function lock(upstream)
-  local _, err = ewma_lock:lock(upstream .. LOCK_KEY)
-  if err then
-    if err ~= "timeout" then
-      core.log.error("EWMA Balancer failed to lock: ", err)
+    local _, err = ewma_lock:lock(upstream .. LOCK_KEY)
+    if err then
+        if err ~= "timeout" then
+            core.log.error("EWMA Balancer failed to lock: ", err)
+        end
     end
-  end
 
-  return err
+    return err
 end
 
 local function unlock()
-  local ok, err = ewma_lock:unlock()
-  if not ok then
-    core.log.error("EWMA Balancer failed to unlock: ", err)
-  end
+    local ok, err = ewma_lock:unlock()
+    if not ok then
+        core.log.error("EWMA Balancer failed to unlock: ", err)
+    end
 
-  return err
+    return err
 end
 
 local function decay_ewma(ewma, last_touched_at, rtt, now)
-  local td = now - last_touched_at
-  td = (td > 0) and td or 0
-  local weight = math.exp(-td/DECAY_TIME)
+    local td = now - last_touched_at
+    td = (td > 0) and td or 0
+    local weight = math.exp(-td / DECAY_TIME)
 
-  ewma = ewma * weight + rtt * (1.0 - weight)
-  return ewma
+    ewma = ewma * weight + rtt * (1.0 - weight)
+    return ewma
 end
 
 local function store_stats(upstream, ewma, now)
-  local success, err, forcible = shm_last_touched_at:set(upstream, now)
-  if not success then
-    core.log.warn("shm_last_touched_at:set failed: ", err)
-  end
-  if forcible then
-    core.log.warn("shm_last_touched_at:set valid items forcibly overwritten")
-  end
+    local success, err, forcible = shm_last_touched_at:set(upstream, now)
+    if not success then
+        core.log.warn("shm_last_touched_at:set failed: ", err)
+    end
+    if forcible then
+        core.log
+            .warn("shm_last_touched_at:set valid items forcibly overwritten")
+    end
 
-  success, err, forcible = shm_ewma:set(upstream, ewma)
-  if not success then
-    core.log.warn("shm_ewma:set failed: ", err)
-  end
-  if forcible then
-    core.log.warn("shm_ewma:set valid items forcibly overwritten")
-  end
+    success, err, forcible = shm_ewma:set(upstream, ewma)
+    if not success then
+        core.log.warn("shm_ewma:set failed: ", err)
+    end
+    if forcible then
+        core.log.warn("shm_ewma:set valid items forcibly overwritten")
+    end
 end
 
 local function get_or_update_ewma(upstream, rtt, update)
-  local lock_err = nil
-  if update then
-    lock_err = lock(upstream)
-  end
+    local lock_err = nil
+    if update then
+        lock_err = lock(upstream)
+    end
 
-  local ewma = shm_ewma:get(upstream) or 0
-  if lock_err ~= nil then
-    return ewma, lock_err
-  end
+    local ewma = shm_ewma:get(upstream) or 0
+    if lock_err ~= nil then
+        return ewma, lock_err
+    end
 
-  local now = ngx_now()
-  local last_touched_at = ngx.shared.balancer_ewma_last_touched_at:get(upstream) or 0
-  ewma = decay_ewma(ewma, last_touched_at, rtt, now)
+    local now = ngx_now()
+    local last_touched_at = ngx.shared.balancer_ewma_last_touched_at:get(
+                                upstream) or 0
+    ewma = decay_ewma(ewma, last_touched_at, rtt, now)
 
-  if not update then
+    if not update then
+        return ewma, nil
+    end
+
+    store_stats(upstream, ewma, now)
+
+    unlock()
+
     return ewma, nil
-  end
-
-  store_stats(upstream, ewma, now)
-
-  unlock()
-
-  return ewma, nil
 end
 
 local function get_upstream_name(upstream)
-   return upstream.host .. ":" .. upstream.port
+    return upstream.host .. ":" .. upstream.port
 end
 
 local function score(upstream)
-  -- Original implementation used names
-  -- Endpoints don't have names, so passing in IP:Port as key instead
-  local upstream_name = get_upstream_name(upstream)
-  return get_or_update_ewma(upstream_name, 0, false)
+    -- Original implementation used names
+    -- Endpoints don't have names, so passing in IP:Port as key instead
+    local upstream_name = get_upstream_name(upstream)
+    return get_or_update_ewma(upstream_name, 0, false)
 end
 
 local function parse_addr(addr)
@@ -126,7 +122,7 @@ end
 
 local function _trans_format(up_nodes)
     -- trans
-    --{"1.2.3.4:80":100,"5.6.7.8:8080":100}
+    -- {"1.2.3.4:80":100,"5.6.7.8:8080":100}
     -- into
     -- [{"host":"1.2.3.4","port":"80"},{"host":"5.6.7.8","port":"8080"}]
     local peers = {}
@@ -147,8 +143,7 @@ end
 local function _ewma_find(ctx, up_nodes)
     local peers
 
-    if not up_nodes
-       or core.table.nkeys(up_nodes) == 0 then
+    if not up_nodes or core.table.nkeys(up_nodes) == 0 then
         return nil, 'up_nodes empty'
     end
 
@@ -161,9 +156,9 @@ local function _ewma_find(ctx, up_nodes)
     local endpoint, backendpoint = peers[1], nil
 
     if #peers > 1 then
-        local a, b = math.random(1, #peers), math.random(1, #peers-1)
+        local a, b = math.random(1, #peers), math.random(1, #peers - 1)
         if b >= a then
-            b = b+1
+            b = b + 1
         end
 
         endpoint, backendpoint = peers[a], peers[b]
@@ -174,7 +169,6 @@ local function _ewma_find(ctx, up_nodes)
 
     return get_upstream_name(endpoint)
 end
-
 
 local function _ewma_after_balance(ctx, before_retry)
     if before_retry then
@@ -195,14 +189,13 @@ local function _ewma_after_balance(ctx, before_retry)
 end
 
 function _M.new(up_nodes, upstream)
-    if not shm_ewma
-       or not shm_last_touched_at then
+    if not shm_ewma or not shm_last_touched_at then
         return nil, "dictionary not find"
     end
 
     return {
         upstream = upstream,
-        get = function (ctx)
+        get = function(ctx)
             return _ewma_find(ctx, up_nodes)
         end,
         after_balance = _ewma_after_balance