--- a/apisix/balancer/ewma.lua
+++ b/apisix/balancer/ewma.lua
@@ -11,6 +11,7 @@ local ngx_shared = ngx.shared
 local ngx_now = ngx.now
 local math = math
 local pairs = pairs
+local next = next
 
 local DECAY_TIME = 10 -- this value is in seconds
 local LOCK_KEY = ":ewma_key"