@hnlq715 2021-01-28
> And we need a test for "fiter tried servers".

@spacewander test case added.